==> analysis/13_fig1/scritps/09_assemble_figs.Rmd <==
---
title: "09 figs"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
```

# Left: Day 5. Right: Day 3.

```{r out.width='50%' }
knitr::include_graphics(c('../../../../13_fig1/figs/FDR005 Splicing type pie chart in SIG events.png','../figs/' ))
knitr::include_graphics(c('../figs/FDR025 Splicing type pie chart in SIG events.png','../figs/type dist of the two.jpg'))

```

```{r}

```
